Calculating time and money based on parental leave and nursery

Calculating time and money based on parental leave and nursery

Usage

babytime(
  parent1_wp,
  parent2_wp,
  baby_waking_time,
  baby_sleeping_time,
  parental_leave_option,
  statutory_weekly_pay = 156.66,
  hourly_nursery_rate,
  parent1_annual_salary,
  parent2_annual_salary,
  annual_salary_increment = 0.02
)

Arguments

parent1_wp

Work plan for parent 1, output from work_plan

parent2_wp

Work plan for parent 2, output from work_plan

baby_waking_time

Expected time that baby wakes up in the morning e.g. "07:00"

baby_sleeping_time

Expected time that baby sleeps at night e.g. "19:00"

parental_leave_option

Choose option 1 or 2 as per http://www.bristol.ac.uk/hr/policies/shared-parental-leave.html

statutory_weekly_pay

Expected weekly pay (as of April 2022 it was 156.66)

hourly_nursery_rate

Hourly nursery cost

parent1_annual_salary

Gross salary of parent 1 from day 1

parent2_annual_salary

Gross salary of parent 2 from day 1

annual_salary_increment

Expected annual increment, default=0.02

Value

List of results

parent1_wp = Parent 1 working times

parent2_wp = Parent 2 working times

baby_schedule = Baby's schedule

overlapping_times = Overlapping times between parents and baby

parent1_salary = Parent 1 salary over time

parent2_salary = Parent 2 salary over time

statutory_parental_pay = Statutory parental pay over time

salary_loss = Salary reductions due to leave

nursery_cost = Nursery costs

weekly_monies = Summary of weekly incomes and outgoings

babytime = Summary of time spent with baby

earnings_plot = Plot of weekly net earnings

babytime_plot = Plot of weekly baby time